+++ forms-flow-service/src/constants/constants.ts
@@ -12,4 +12,10 @@ export const MULTITENANCY_ENABLED =
 
 //application details
 export const APPLICATION_NAME =
-        (window._env_ && window._env_.REACT_APP_APPLICATION_NAME) || "formsflow.ai";
+        (window as any).REACT_APP_APPLICATION_NAME || "roadsafety"
+        // "formsflow.ai";
+
+// Used in encyrpting and decrypting the token from local storage.
+export const TOKEN_ENCRYPTION_KEY =
+        (window as any).REACT_APP_TOKEN_ENCRYPTION_KEY ||
+        "8f4a4e01b639aa73d2b5bdb6e9f2e6aef471b3dbba3d5e48e3a798fc3d6d6cbb";
