+++ forms-flow-service/src/helpers/helperServices.ts
@@ -1,5 +1,6 @@
 import moment from 'moment'
-import { DATE_FORMAT ,MULTITENANCY_ENABLED ,TIME_FORMAT} from '../constants/constants';
+import CryptoJS from "crypto-js";
+import { DATE_FORMAT ,MULTITENANCY_ENABLED ,TIME_FORMAT, TOKEN_ENCRYPTION_KEY} from '../constants/constants';
 
 class HelperServices {
   public static getLocalDateAndTime(date: string): any {
@@ -84,6 +85,50 @@ class HelperServices {
       exactRouteMatches.some((route) => location == route) || 
       partOfRouteMatches.some((route) => location.includes(route))
     );
+  }
+
+  /**
+   * Encrypts a given text using AES encryption.
+   * @param text - The text to encrypt (e.g., refresh token).
+   * @returns Encrypted string (Base64 format).
+   * @throws Error if encryption fails.
+   */
+  public static encrypt(text: string): string {
+    try {
+      if (!text) {
+        throw new Error("Encryption failed: Input text is empty or undefined.");
+      }
+      return CryptoJS.AES.encrypt(text, TOKEN_ENCRYPTION_KEY).toString();
+    } catch (error) {
+      console.error("Encryption error:", error);
+      throw new Error("Encryption failed. See console for details.");
+    }
+  }
+
+  /**
+   * Decrypts an AES-encrypted string.
+   * @param encryptedText - The encrypted string to decrypt.
+   * @returns The decrypted original string.
+   * @throws Error if decryption fails.
+   */
+  public static decrypt(encryptedText: string): string {
+    try {
+      if (!encryptedText) {
+        throw new Error("Decryption failed: Input is empty or undefined.");
+      }
+
+      const bytes = CryptoJS.AES.decrypt(encryptedText, TOKEN_ENCRYPTION_KEY);
+      const decryptedText = bytes.toString(CryptoJS.enc.Utf8);
+
+      if (!decryptedText) {
+        throw new Error("Decryption failed: Invalid ciphertext or wrong key.");
+      }
+
+      return decryptedText;
+    } catch (error) {
+      console.error("Decryption error:", error);
+      throw new Error("Decryption failed. See console for details.");
+    }
   }
 }
